requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pillow
matplotlib
pandas
img2pdf
python-dotenv